+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2024-10-18] [add event dispatcher to webhooks](https://github.com/RubricLab/webhooks/commit/069e76931da1bd49f05ab27cd6119ffec94a0c8c)
 - [2024-10-03] [mod lint script](https://github.com/RubricLab/webhooks/commit/ebf8bf9d4ededf1abfb6ff88e0c34142af03f869)
 - [2024-10-03] [gitignore turbo, format](https://github.com/RubricLab/webhooks/commit/607dfed6ed945bb02ff9fb56e6755c6b826c5ddc)
 - [2024-10-02] [init](https://github.com/RubricLab/webhooks/commit/fa8f2f83d7cbc48865f343e7af69fd7fa8ec2d37)